litfinder/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Word document generation
python-docx==1.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# LLM
anthropic==0.12.0

# Telegram
aiogram==3.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Export formats
python-docx==1.1.0
bibtexparser==1.4.1

# XML Parsing (OAI-PMH)
lxml==5.1.0
xmltodict==0.13.0

# Utils
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
